Creacion de herencia Sala
