+++ src/com/IesAlonsoMadrigal/domain/models/Sala_Ordenadores.java
@@ -1,28 +1,12 @@
 package com.IesAlonsoMadrigal.domain.models;
 
-public class Sala_Ordenadores {
+public class Sala_Ordenadores extends Sala{
 
-    private Integer id;
-    private String planta;
+
     private String num_ordenadores;
-    private String capacidad;
     private String tipo_ordenadores;
 
-    public Integer getId() {
-        return id;
-    }
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getPlanta() {
-        return planta;
-    }
-
-    public void setPlanta(String planta) {
-        this.planta = planta;
-    }
 
     public String getNum_ordenadores() {
         return num_ordenadores;
@@ -32,13 +16,6 @@ public class Sala_Ordenadores {
         this.num_ordenadores = num_ordenadores;
     }
 
-    public String getCapacidad() {
-        return capacidad;
-    }
-
-    public void setCapacidad(String capacidad) {
-        this.capacidad = capacidad;
-    }
 
     public String getTipo_ordenadores() {
         return tipo_ordenadores;
