--- src/com/IesAlonsoMadrigal/domain/models/Salas.java
@@ -1,5 +1,0 @@
-package com.IesAlonsoMadrigal.domain.models;
-
-public class Salas {
-    Hola
-}
